+++ lib/webpacker/file_loader.rb
@@ -8,15 +8,11 @@ class Webpacker::FileLoader
 
   class << self
     def exist?
-      instance.data.present? rescue false
+      instance.data.try(&:present?)
     end
 
     def load(path = file_path)
       self.instance = new(path)
-    end
-
-    def reloadable?
-      ["development", "test"].include?(Rails.env)
     end
   end